refactor SDK for 0.9.5

+++ CausalFoundrySDKEcom/Sources/Models/ItemObjects/StoreObject.swift
@@ -9,8 +9,8 @@ import Foundation
 
 public struct StoreObject: Codable {
     var id: String
-    var lat: Float? = 0
-    var lon: Float? = 0
+    var lat: Double? = 0
+    var lon: Double? = 0
 
     // Custom coding keys enum
     enum CodingKeys: String, CodingKey {
@@ -19,7 +19,7 @@ public struct StoreObject: Codable {
         case lon
     }
 
-    public init(id: String, lat: Float? = 0, lon: Float? = 0) {
+    public init(id: String, lat: Double? = 0, lon: Double? = 0) {
         self.id = id
         self.lat = lat
         self.lon = lon
@@ -37,7 +37,7 @@ public struct StoreObject: Codable {
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         id = try container.decode(String.self, forKey: .id)
-        lat = try container.decode(Float.self, forKey: .lat)
-        lon = try container.decode(Float.self, forKey: .lon)
+        lat = try container.decode(Double.self, forKey: .lat)
+        lon = try container.decode(Double.self, forKey: .lon)
     }
 }
